refactor: 使用 Lombok 的 @RequiredArgsConstructor 注解
- 移除了多个文件中的 @Autowired 注解和对应的构造函数
- 在类级别添加了 @RequiredArgsConstructor 注解，自动生成构造函数
-简化了代码结构，提高了可读性和维护性
