+++ chuang-api/src/main/java/cn/zhangchuangla/api/controller/system/RegisterController.java
@@ -7,6 +7,7 @@ import cn.zhangchuangla.infrastructure.web.service.RegisterService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -21,13 +22,10 @@ import org.springframework.web.bind.annotation.RestController;
 @Slf4j
 @RestController
 @Tag(name = "注册接口")
+@RequiredArgsConstructor
 public class RegisterController extends BaseController {
 
     private final RegisterService registerService;
-
-    public RegisterController(RegisterService registerService) {
-        this.registerService = registerService;
-    }
 
 
     /**
