+++ chuang-api/src/main/java/cn/zhangchuangla/api/controller/system/SysDeptController.java
@@ -16,8 +16,8 @@ import cn.zhangchuangla.system.service.SysDeptService;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.BeanUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PostAuthorize;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
@@ -34,14 +34,10 @@ import java.util.List;
 @Tag(name = "部门管理")
 @RestController
 @RequestMapping("/system/department")
+@RequiredArgsConstructor
 public class SysDeptController extends BaseController {
 
     private final SysDeptService sysDeptService;
-
-    @Autowired
-    public SysDeptController(SysDeptService sysDeptService) {
-        this.sysDeptService = sysDeptService;
-    }
 
 
     /**
