+++ chuang-api/src/main/java/cn/zhangchuangla/api/controller/system/SysDictionaryDataController.java
@@ -17,8 +17,8 @@ import cn.zhangchuangla.system.service.DictionaryDataService;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.BeanUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -34,14 +34,10 @@ import java.util.List;
 @RestController
 @Anonymous
 @RequestMapping("/system/dictionary/data")
+@RequiredArgsConstructor
 public class SysDictionaryDataController extends BaseController {
 
     private final DictionaryDataService dictionaryDataService;
-
-    @Autowired
-    public SysDictionaryDataController(DictionaryDataService dictionaryDataService) {
-        this.dictionaryDataService = dictionaryDataService;
-    }
 
 
     /**
