+++ chuang-api/src/main/java/cn/zhangchuangla/api/controller/system/SysPermissionsController.java
@@ -9,6 +9,7 @@ import cn.zhangchuangla.system.service.SysPermissionsService;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -24,14 +25,11 @@ import java.util.List;
 @RequestMapping
 @Tag(name = "权限接口")
 @RestController("/system/permission")
+@RequiredArgsConstructor
 public class SysPermissionsController extends BaseController {
 
 
     private final SysPermissionsService sysPermissionsService;
-
-    public SysPermissionsController(SysPermissionsService sysPermissionsService) {
-        this.sysPermissionsService = sysPermissionsService;
-    }
 
 
     /**
