+++ chuang-api/src/main/java/cn/zhangchuangla/api/controller/system/SysRoleController.java
@@ -6,7 +6,6 @@ import cn.zhangchuangla.common.core.page.TableDataResult;
 import cn.zhangchuangla.common.enums.BusinessType;
 import cn.zhangchuangla.common.enums.ResponseCode;
 import cn.zhangchuangla.common.result.AjaxResult;
-import cn.zhangchuangla.infrastructure.annotation.Anonymous;
 import cn.zhangchuangla.infrastructure.annotation.OperationLog;
 import cn.zhangchuangla.system.model.entity.SysRole;
 import cn.zhangchuangla.system.model.request.role.SysRoleAddRequest;
@@ -17,9 +16,9 @@ import cn.zhangchuangla.system.service.SysRoleService;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -37,15 +36,10 @@ import java.util.List;
 @RestController
 @RequestMapping("/system/role")
 @Tag(name = "角色接口")
-@Anonymous
+@RequiredArgsConstructor
 public class SysRoleController extends BaseController {
 
     private final SysRoleService sysRoleService;
-
-    @Autowired
-    public SysRoleController(SysRoleService sysRoleService) {
-        this.sysRoleService = sysRoleService;
-    }
 
 
     /**
