+++ chuang-system/chuang-system-core/src/main/java/cn/zhangchuangla/system/service/impl/SysPostServiceImpl.java
@@ -10,8 +10,8 @@ import cn.zhangchuangla.system.service.SysPostService;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.BeanUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,15 +22,12 @@ import java.util.List;
  * @author zhangchuang
  */
 @Service
+@RequiredArgsConstructor
 public class SysPostServiceImpl extends ServiceImpl<SysPostMapper, SysPost>
         implements SysPostService {
 
     private final SysPostMapper sysPostMapper;
 
-    @Autowired
-    public SysPostServiceImpl(SysPostMapper sysPostMapper) {
-        this.sysPostMapper = sysPostMapper;
-    }
 
     /**
      * 岗位列表
