+++ chuang-system/chuang-system-storage/src/main/java/cn/zhangchuangla/storage/loader/SysFileConfigLoader.java
@@ -11,8 +11,8 @@ import cn.zhangchuangla.storage.model.entity.SysFileConfig;
 import cn.zhangchuangla.storage.service.SysFileConfigService;
 import com.alibaba.fastjson.JSON;
 import jakarta.annotation.PostConstruct;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.Map;
@@ -24,17 +24,12 @@ import java.util.concurrent.ConcurrentHashMap;
  */
 @Component
 @Slf4j
+@RequiredArgsConstructor
 public class SysFileConfigLoader {
 
     private final SysFileConfigService sysFileConfigService;
     private final AppProperty appProperty;
     private final Map<String, String> sysFileConfigCache = new ConcurrentHashMap<>(4);
-
-    @Autowired
-    public SysFileConfigLoader(SysFileConfigService sysFileConfigService, AppProperty appProperty) {
-        this.sysFileConfigService = sysFileConfigService;
-        this.appProperty = appProperty;
-    }
 
     /**
      * 系统启动时加载配置
